Refactor prompt handling and update collections structure in documentation
