+++ mkdocs.yml
@@ -52,7 +52,9 @@ nav:
     - Create Summarization Analysis: tutorials/create_summarization.md
     - Create Custom Services: tutorials/create_custom_services.md
   - Database: database.md
-  - Collections: collections.md
+  - Collections:
+    - Overview: collections.md
+    - Cmetadata: cmetadata.md
   # /index/... endpoints
   - Document Index: document_index.md
   # /chat, /search, /chat_history
